requirements.txt
streamlit
openai-whisper
ffmpeg-python
faiss-cpu
sentence-transformers
numpy
